insert logic for create and delete delegation with correct builder of data and refactor how to use yaml file
