+++ src/main/java/it/pn/frontend/e2e/model/delegate/Delegator.java
@@ -1,4 +1,4 @@
-package it.pn.frontend.e2e.model;
+package it.pn.frontend.e2e.model.delegate;
 
 import lombok.Builder;
 import lombok.Data;
