+++ src/test/resources/dataPopulation/nuovaDelegaPG.yaml
@@ -1,4 +1,4 @@
 ragioneSociale: "Le Epistolae srl"
 codiceFiscale: "LELPTR04A01C352E"
 ente: "Comune di Palermo"
-codiceDelega: "24411"
+codiceDelega: "41992"
