+++ src/test/resources/feature/3-destinatario/personaGiuridica/deleghePG/003_114_visualizzaDelegheNonAssegnataGruppoPG.feature
@@ -18,7 +18,4 @@ Feature: La persona giuridica visualizza le deleghe
     And Nella pagina Deleghe si clicca su Delegati dall impresa
     And Si visualizza correttamente la pagina Deleghe sezione Deleghe dell impresa
     And Nella pagina Deleghe si clicca su Deleghe a carico dell impresa
-    And Si visualizza correttamente la pagina Deleghe sezione Deleghe a Carico dell impresa
-    And Si vede correttamente l'elenco delle deleghe
-    And Si ripristina lo stato iniziale delle deleghe a carico dell impresa "personaGiuridica"
     And Logout da portale persona giuridica
