+++ src/test/resources/feature/3-destinatario/personaGiuridica/deleghePG/004_119_accettazioneDelegaSenzaGruppo.feature
@@ -5,23 +5,16 @@ Feature:Il delegato persona giuridica accetta la delega non assegnandoli un grup
     @DeleghePG
     @PG
 
-  Scenario Outline: PN-9171 - Il delegato persona giuridica accetta la delega non assegnandoli un gruppo
+  Scenario: PN-9171 - Il delegato persona giuridica accetta la delega non assegnandoli un gruppo
     Given PG - Si effettua la login tramite token exchange come "delegante", e viene visualizzata la dashboard
     When Nella pagina Piattaforma Notifiche persona giuridica click sul bottone Deleghe
     And Si visualizza correttamente la pagina Deleghe sezione Deleghe a Carico dell impresa
     And Nella pagina Deleghe sezione Deleghe a carico dell'impresa si controlla la presenza di una delega per PG
-      | accessoCome    | delegante     |
+      | accessoCome    | delegato      |
       | fiscalCode     | 27957814470   |
       | companyName    | Convivio Spa  |
       | displayName    | Convivio Spa  |
       | person         | false         |
-    And Si sceglie opzione accetta
-    And Si inserisce il codice della delega a carico dell impresa nella modale
-    And Si clicca sul bottone conferma delega
-    And Si clicca su conferma in assegnazione gruppo
-    And Si controlla che la delega PG ha lo stato Attiva <ragioneSociale>
-   # And Si ripristina lo stato iniziale delle deleghe a carico dell impresa "personaGiuridica"
+    And Si accetta la delega "senza" gruppo
+    And Si controlla che la delega PG ha lo stato Attiva "Le Epistolae srl"
     And Logout da portale persona giuridica
-    Examples:
-      | ragioneSociale |
-      | Convivio Spa   |
