+++ src/test/resources/feature/3-destinatario/personaGiuridica/deleghePG/008_115_ricercaDelegheConGruppoPG.feature
@@ -16,13 +16,8 @@ Feature:La persona giuridica fa una ricerca per gruppo delle deleghe
       | companyName    | Convivio Spa  |
       | displayName    | Convivio Spa  |
       | person         | false         |
-    And Si sceglie opzione accetta
-    And Si inserisce il codice della delega a carico dell impresa nella modale
-    And Nella sezione Deleghe si clicca sul bottone conferma codice
-    And Si assegna un gruppo alla delega
-    And Si clicca sul bottone conferma gruppo
+    And Si accetta la delega "con" gruppo
     And Nella pagina Deleghe sezione Deleghe a Carico dell impresa si inserisce il gruppo del delegante
     And Nella pagina Deleghe sezione Deleghe a Carico dell impresa si clicca su bottone Filtra
     And Nella pagina Deleghe sezione Deleghe a Carico dell impresa si controlla che venga restituita la delega con il codice fiscale inserito "personaGiuridica_1"
-    And Si ripristina lo stato iniziale delle deleghe a carico dell impresa "personaGiuridica"
     And Logout da portale persona giuridica
