+++ src/test/resources/feature/3-destinatario/personaGiuridica/deleghePG/011_121_1_modificaDelegaSenzaGruppo.feature
@@ -15,11 +15,10 @@ Feature: Il delegato persona giuridica modifica una delega non assegnandoli un g
       | companyName    | Convivio Spa  |
       | displayName    | Convivio Spa  |
       | person         | false         |
-    And Si accetta la delega con un gruppo
+    And Si accetta la delega "con" gruppo
     And Nella pagina Deleghe sezione Deleghe a carico dell'impresa clicca sul menu della delega "personaGiuridica"
     And Nella sezione Deleghe si clicca sul bottone modifica
     And Si clicca sul bottone non assegna a un gruppo
     And Si clicca su conferma in assegnazione gruppo
     And Si controlla che la delega non abbia più il gruppo
-    And Si ripristina lo stato iniziale delle deleghe a carico dell impresa "personaGiuridica"
     And Logout da portale persona giuridica
